[Refactor] useUpdateBeforeHeight hook으로 분리
